Add extensions for generic services
